+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* **2013-07-26**: [Model] Deprecated setRouteContent and getRouteContent, use setContent and getContent instead.
 * **2013-07-19**: [Model] Separated database agnostic, doctrine generic and
   PHPCR-ODM specific code to prepare for Doctrine ORM support.
 * **2013-07-17**: [FormType] Moved TermsFormType to CoreBundle and renamed it to CheckboxUrlLableFormType